requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
google-genai==1.29.0
pydantic>=2.9,<3